disable on mimxrt10xx (Teensy40 41) Serial port Hooks
As mentioned in issue #6241  the commit to setup port hooks is now
causing all input/output that are to go to the Mu window to also
go to the LpUart that is defined the port serial.c

and in this case it goes to lpuart4, which on Teensy 4, 4.1 is
used on Arduino Serial2.  With this new code this port no longer
works properly.

This is one way to solve it, in that there is a #if defined() that
if not set, all of the code in this file is ignored and the higher
level supervisor stub versions of these functions will be used, which
don't interfere with Serial2 and my test sketch works again.

Note: the PR for Switch to Port Serial Hooks, also changed code in
other ports.  I have not tried to see how.

There are other more global fixes for this, in which maybe a higer
level  #if that disables the code within the top level supervisor.  Or
could be software controlled

Again this may not be the final solution, but at least it gets
Serial2 up and running agin.